Altered specific variable data types from int to double for both Problem1 and Problem2 programs, which produced expected output results.

+++ src/edu/dmacc/coma510/Problem1.java
@@ -1,12 +1,14 @@
 package edu.dmacc.coma510;
 
+/* DPK - 10/12/2019 - changed numberOfPlateAppearances and calculatedBattingAverage variables from int to double data type */
+
 public class Problem1 {
     public static void main(String[] args) {
-        int numberOfPlateAppearances = 100;
+        double numberOfPlateAppearances = 100;
         int numberOfWalks = 4;
         int numberOfHits = 24;
 
-        int calculatedBattingAverage = numberOfHits / (numberOfPlateAppearances - numberOfWalks);
+        double calculatedBattingAverage = numberOfHits / (numberOfPlateAppearances - numberOfWalks);
 
         System.out.println(calculatedBattingAverage); // 0.25
     }
